[10.x] Refactor shared static `methodExcludedByOptions` method to trait (#46498)
* Refactor shared methodExcludedByOptions static method to trait

* formatting

* fix test

---------
