+++ src/Illuminate/Routing/ControllerDispatcher.php
@@ -7,7 +7,7 @@ use Illuminate\Routing\Contracts\ControllerDispatcher as ControllerDispatcherCon
 
 class ControllerDispatcher implements ControllerDispatcherContract
 {
-    use RouteDependencyResolverTrait;
+    use FiltersControllerMiddleware, ResolvesRouteDependencies;
 
     /**
      * The container instance.
@@ -77,18 +77,5 @@ class ControllerDispatcher implements ControllerDispatcherContract
         return collect($controller->getMiddleware())->reject(function ($data) use ($method) {
             return static::methodExcludedByOptions($method, $data['options']);
         })->pluck('middleware')->all();
-    }
-
-    /**
-     * Determine if the given options exclude a particular method.
-     *
-     * @param  string  $method
-     * @param  array  $options
-     * @return bool
-     */
-    public static function methodExcludedByOptions($method, array $options)
-    {
-        return (isset($options['only']) && ! in_array($method, (array) $options['only'])) ||
-            (! empty($options['except']) && in_array($method, (array) $options['except']));
     }
 }
